LU-13005 lnet: Support checking for MD leaks.
Since we dropped the refcounting on LNetEQ we no longer get
confirmation that all MDs for a given handler a gone by the
time they should be.

So add lnet_assert_handler_unused() which searches the per-cpt
containers and ensures there are no MDs for a given handler, and call
that are the same place that we used to call LNetEQFree().

Signed-off-by: Mr NeilBrown <neilb@suse.de>
Change-Id: Ic04e5b4f67a79055c49994bd1d92e275fa71dd4b
Reviewed-on: https://review.whamcloud.com/38059
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Chris Horn <chris.horn@hpe.com>
Reviewed-by: James Simmons <jsimmons@infradead.org>
Reviewed-by: Oleg Drokin <green@whamcloud.com>