LU-13301 lfs: return error when setstripe -o on inactive OSTs
When running "lfs setstripe" on inactive OSTs with -o option,
error code ENOTCONN returned by osp_statfs() for OST inactive
state is mistreated as MDT disconnection by the client, then the
client keeps trying to restore the connection to MDT and resend
the request again and again.
To fix this issue, the error code "ENOTCONN" is changed to other
code (EREMOTEIO) for distinction.

sanity.sh test_27cf is added to verify this patch.

Signed-off-by: Emoly Liu <emoly@whamcloud.com>
Change-Id: I9ea10b4985c769e52ee54ff22ac69948073b7709
Reviewed-on: https://review.whamcloud.com/37751
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Andreas Dilger <adilger@whamcloud.com>
Reviewed-by: Li Xi <lixi@ddn.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>